Moved dosage settings to data folder and renamed the data models file
